requirements.txt
Flask
faiss-cpu
numpy
google-generativeai
